Adjust specs and move some assessments related specs to the right folder
