+++ packages/yoastseo/spec/scoring/assessorSpec.js
@@ -1,9 +1,9 @@
-import DefaultResearcher from "../src/languageProcessing/languages/_default/Researcher";
-import Assessor from "../src/scoring/assessor.js";
-import Paper from "../src/values/Paper.js";
-import AssessmentResult from "../src/values/AssessmentResult.js";
-import MissingArgument from "../src/errors/missingArgument";
-import factory from "./specHelpers/factory.js";
+import DefaultResearcher from "../../src/languageProcessing/languages/_default/Researcher";
+import Assessor from "../../src/scoring/assessor.js";
+import Paper from "../../src/values/Paper.js";
+import AssessmentResult from "../../src/values/AssessmentResult.js";
+import MissingArgument from "../../src/errors/missingArgument";
+import factory from "../specHelpers/factory.js";
 var i18n = factory.buildJed();
 
 global.window = {};
