+++ packages/yoastseo/spec/scoring/cornerstone/relatedKeywordAssessorSpec.js
@@ -1,8 +1,8 @@
-import EnglishResearcher from "../../src/languageProcessing/languages/en/Researcher";
-import Assessor from "../../src/scoring/cornerstone/relatedKeywordAssessor.js";
-import Paper from "../../src/values/Paper.js";
-import factory from "../specHelpers/factory.js";
-import getResults from "../specHelpers/getListOfAssessmentResults";
+import EnglishResearcher from "../../../src/languageProcessing/languages/en/Researcher";
+import Assessor from "../../../src/scoring/cornerstone/relatedKeywordAssessor.js";
+import Paper from "../../../src/values/Paper.js";
+import factory from "../../specHelpers/factory.js";
+import getResults from "../../specHelpers/getListOfAssessmentResults";
 const i18n = factory.buildJed();
 const assessor = new Assessor( i18n, { researcher: new EnglishResearcher() } );
