+++ packages/yoastseo/spec/scoring/relatedKeywordTaxonomyAssessorSpec.js
@@ -1,10 +1,10 @@
-import EnglishResearcher from "../src/languageProcessing/languages/en/Researcher";
-import Assessor from "../src/scoring/relatedKeywordTaxonomyAssessor.js";
-import Paper from "../src/values/Paper.js";
-import factory from "./specHelpers/factory.js";
+import EnglishResearcher from "../../src/languageProcessing/languages/en/Researcher";
+import Assessor from "../../src/scoring/relatedKeywordTaxonomyAssessor.js";
+import Paper from "../../src/values/Paper.js";
+import factory from "../specHelpers/factory.js";
 const i18n = factory.buildJed();
 const assessor = new Assessor( i18n, { researcher: new EnglishResearcher() } );
-import getResults from "./specHelpers/getAssessorResults";
+import getResults from "../specHelpers/getAssessorResults";
 
 describe( "running assessments in the assessor", function() {
 	it( "runs assessments without any specific requirements", function() {
