+++ packages/yoastseo/spec/scoring/taxonomyAssessorSpec.js
@@ -1,9 +1,9 @@
-import DefaultResearcher from "../src/languageProcessing/languages/_default/Researcher";
-import Assessor from "../src/scoring/taxonomyAssessor.js";
-import Paper from "../src/values/Paper.js";
-import factory from "./specHelpers/factory.js";
+import DefaultResearcher from "../../src/languageProcessing/languages/_default/Researcher";
+import Assessor from "../../src/scoring/taxonomyAssessor.js";
+import Paper from "../../src/values/Paper.js";
+import factory from "../specHelpers/factory.js";
 const i18n = factory.buildJed();
-import getResults from "./specHelpers/getAssessorResults";
+import getResults from "../specHelpers/getAssessorResults";
 
 describe( "running assessments in the assessor", function() {
 	let assessor;
